Fix int8_t overflow in AAF frequency selection for ICM-42688P/42686P
selectedFreq in getGyroAafConfig() was declared int8_t (max 127), but
aafLUT42688[] contains freq values up to 1962. Values above 127 silently
truncated on assignment, corrupting the closest-frequency comparison and
causing the wrong AAF register configuration to be selected. This affected
every board using this gyro at the default 256Hz LPF setting (selected
303Hz instead of the correct 258Hz cutoff). Changed to uint16_t to match
aafConfig_t.freq's own type.

+++ src/main/drivers/accgyro/accgyro_icm42605.c
@@ -422,7 +422,7 @@ static const aafConfig_t *getGyroAafConfig(icm42605Variant_e variant, const uint
         aafConfigs = aafLUT42605;
     }
     int i;
-    int8_t selectedFreq = aafConfigs[0].freq;
+    uint16_t selectedFreq = aafConfigs[0].freq;
     const aafConfig_t * candidate = &aafConfigs[0];
 
     // Choose closest supported LPF value
